ngclient: Do not use urljoin to form metadata URL
urljoin considers the second URL to override the base URL if the second
one contains e.g. hostname: this could lead to ngclient downloading
from the wrong host entirely. Doing that would not compromise the
security of the system as the metadata would still need to be verified,
but would definitely be unexpected and a bug.

Note that we're still not encoding the rolename, it's just inserted into
the URL as is.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>