ngclient: remove sleep_before_round
The purpose of this config was to ensure blocking
inside the download loop and releasing CPU resources.

To our best knowledge the network stack currently used
in RequestsFetcher will always block inside the loop
and the issue cannot be reproduced.

'chunk_size' and 'socket_timeout' are currently the
settings provided by RequestsFetcher to tweak
CPU usage and download granularity.

Signed-off-by: Teodora Sechkova <tsechkova@vmware.com>