Linting tests with black and exclude old tests
Currently, we are using 4 tools: black, pylint, isort and mypy.
We want to run all 4 of them on our tests for the new codebase to make
it more readable and maintainable, but we don't want to run it on the
old test files as they are following the old style guidelines.

In order to achieve that we want to use an exclusion list instead of
an inclusion list as the new code base will be developed in the future
and new test files will appear. On another hand, we don't expect any
more additional test files against the old code, so this list is static.

I decided to hardcode the names we want to exclude because that way we
won't have to remove the Git history, as opposed to renaming or moving
the old test files.
Even though the list is big around 30 files, I think this solution is
fine as this list will only contain files testing the old code meaning
the list will have static content.

I will apply each of the linters in a separate pr.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>