Linting tests with pylint and exclude old tests
Currently, we are using 4 tools: black, pylint, isort and mypy.
We want to run all 4 of them on our tests for the new codebase to make
it more readable and maintainable, but we don't want to run it on the
old test files as they are following the old style guidelines.

I am using the same approach as the one used for black - create an
exclusion list from old test files we want to ignore.
I didn't used moved or renamed the old test files as that will remove
the Git history.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>