Apply black and isort on tuf/exceptions.py
We are already using tuf/exceptions.py inside
- tuf/api/metadata.py
- tuf/ngclient/__internal__/trusted_metadata_set.py
- tuf/ngclient/fetcher.py
- tuf/ngclient/updater.py
and in our tests that this means we are most likely going to use that
module after we remove the old code.

It make sense to lint it with the other three
linters besides mypy (we already lint tuf/exceptions.py with mypy).

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>