DelegatedRole: support succinct_hash_delegations
Start supporting succinct_hash_delegations attribute inside
DelegatedRole. This change is part of TAP 15 implementation.

Unfortunatly, this change is not compatible with all previos python-tuf
versions as the "name" attribute is described as optional in TAP 15,
meaning it has to be replaced to become the fourth argument
instead of being the first when creating a DelegatedRole instance.

Additionally, serialization tests covering the new use cases as TAP 15
states that only one of "paths", "path_hash_prefixes" and
"succinct_hash_delegations" is allowed in a DelegatedRole instance.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>