Tests delegation graphs: make paths an optional
Inside test_updater_delegation_graphs.py we define the TestDelegation
class. There the "paths" attribute is defined as a "List[str]" and
it has a default value of a "[*]". This is often used so that all
targets are considered as delegated by the delegated role that will
consequently be created with a paths attribute with a value of "[*]".
Still, there are a couple of test cases where we want to give a "None"
value to the "paths" attribute and mypy complains that we can't assign
a "None" value when the expected type is a list of strings.
That's why we have to make "paths" optional.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>