Add SuccinctHashDelegations.find_bin()
Add a public method to calculate and receive the bin name when succinct
hash delegation is used.
It's logical that all of the logic required to calculate the bin name
should be inside the class responsible for that SuccinctHashDelegations.

This will eliminate the need of
SuccinctHashDelegation._find_bin_for_bits() and its test.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>