Add SuccinctHashDelegations.get_bin_name()
Add a SuccinctHashDelegations.get_bin_name() API call which returns the
bin name of a given sequential bin number.
This is useful to easily get the name of the first bin which can be
used as a unique identifier for delegated roles using a succinct hash
delegations.

A unique identifier for such delegated roles with succinct hash
is required by our implementation of Delegations.roles where the keys
are DelegatedRole.name attributes and values are DelegatedRole objects
and the fact that when succinct hash delegations are used
DelegatedRole.name should be populated with "None".

Solves: https://github.com/theupdateframework/python-tuf/issues/1943

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>